Move ProjectFileDescriptor nested classes into separate files
