+++ main/src/addins/MonoDevelop.DesignerSupport/MonoDevelop.DesignerSupport.Projects/ProjectFileDescriptor.StandardStringsConverter.cs
@@ -1,0 +1,74 @@
+﻿//
+//Copyright (c) 2007 Novell, Inc (http://www.novell.com)
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+// THE SOFTWARE.
+//
+//
+
+using System;
+using System.ComponentModel;
+
+namespace MonoDevelop.DesignerSupport
+{
+	partial class ProjectFileDescriptor
+	{
+		abstract class StandardStringsConverter : TypeConverter
+		{
+			public override bool GetStandardValuesSupported (ITypeDescriptorContext context)
+			{
+				return true;
+			}
+
+			public override bool CanConvertFrom (ITypeDescriptorContext context, Type sourceType)
+			{
+				return sourceType == typeof (string) || base.CanConvertFrom (context, sourceType);
+			}
+
+			public override bool CanConvertTo (ITypeDescriptorContext context, Type destinationType)
+			{
+				return destinationType == typeof (string) || base.CanConvertTo (context, destinationType);
+			}
+
+			public override object ConvertFrom (ITypeDescriptorContext context,
+												System.Globalization.CultureInfo culture, object value)
+			{
+				if (value != null && value is string)
+					return value;
+				else
+					return base.ConvertFrom (context, culture, value);
+			}
+
+			public override object ConvertTo (ITypeDescriptorContext context, System.Globalization.CultureInfo culture,
+											  object value, Type destinationType)
+			{
+				if (value != null && (destinationType == typeof (string)))
+					return value;
+				else
+					return base.ConvertTo (context, culture, value, destinationType);
+			}
+
+			public override StandardValuesCollection GetStandardValues (ITypeDescriptorContext context)
+			{
+				return new StandardValuesCollection (GetStandardStrings (context));
+			}
+
+			public abstract System.Collections.ICollection GetStandardStrings (ITypeDescriptorContext context);
+		}
+	}
+}
