+++ main/src/addins/MonoDevelop.DesignerSupport/MonoDevelop.DesignerSupport.Projects/ProjectFileDescriptor.cs
@@ -1,4 +1,4 @@
-﻿// ProjectFileDescriptor.cs
+// ProjectFileDescriptor.cs
 //
 //Author:
 //  Lluis Sanchez Gual
@@ -35,7 +35,7 @@ using System.Linq;
 
 namespace MonoDevelop.DesignerSupport
 {
-	class ProjectFileDescriptor: CustomDescriptor, IDisposable
+	partial class ProjectFileDescriptor : CustomDescriptor, IDisposable
 	{
 		ProjectFile file;
 		Project project;
@@ -70,31 +70,31 @@ namespace MonoDevelop.DesignerSupport
 				project = null;
 			}
 		}
-		
+
 		[LocalizedCategory ("Misc")]
 		[LocalizedDisplayName ("Name")]
 		[LocalizedDescription ("Name of the file.")]
 		public string Name {
 			get { return System.IO.Path.GetFileName (file.Name); }
 		}
-		
+
 		[LocalizedCategory ("Misc")]
 		[LocalizedDisplayName ("Path")]
 		[LocalizedDescription ("Full path of the file.")]
 		public string Path {
 			get { return file.FilePath; }
 		}
-		
+
 		[LocalizedCategory ("Misc")]
 		[LocalizedDisplayName ("Type")]
 		[LocalizedDescription ("Type of the file.")]
 		public string FileType {
 			get {
 				string type = IdeServices.DesktopService.GetMimeTypeForUri (file.Name);
-				return IdeServices.DesktopService.GetMimeTypeDescription (type); 
+				return IdeServices.DesktopService.GetMimeTypeDescription (type);
 			}
 		}
-		
+
 		[LocalizedCategory ("Build")]
 		[LocalizedDisplayName ("Build action")]
 		[LocalizedDescription ("Action to perform when building this file.")]
@@ -104,7 +104,7 @@ namespace MonoDevelop.DesignerSupport
 			get { return file.BuildAction; }
 			set { file.BuildAction = value; }
 		}
-		
+
 		[LocalizedCategory ("Build")]
 		[LocalizedDisplayName ("Resource ID")]
 		[LocalizedDescription ("Identifier of the embedded resource.")]
@@ -112,7 +112,7 @@ namespace MonoDevelop.DesignerSupport
 			get { return file.ResourceId; }
 			set { file.ResourceId = value; }
 		}
-		
+
 		[LocalizedCategory ("Build")]
 		[LocalizedDisplayName ("Copy to output directory")]
 		[LocalizedDescription ("Whether to copy the file to the project's output directory when the project is built.")]
@@ -120,7 +120,7 @@ namespace MonoDevelop.DesignerSupport
 			get { return file.CopyToOutputDirectory; }
 			set { file.CopyToOutputDirectory = value; }
 		}
-		
+
 		[LocalizedCategory ("Build")]
 		[LocalizedDisplayName ("Custom Tool")]
 		[LocalizedDescription ("The ID of a custom code generator.")]
@@ -128,7 +128,7 @@ namespace MonoDevelop.DesignerSupport
 			get { return file.Generator; }
 			set { file.Generator = value; }
 		}
-		
+
 		[LocalizedCategory ("Build")]
 		[LocalizedDisplayName ("Custom Tool Namespace")]
 		[LocalizedDescription ("Overrides the namespace in which the custom code generator should generate code.")]
@@ -136,117 +136,10 @@ namespace MonoDevelop.DesignerSupport
 			get { return file.CustomToolNamespace; }
 			set { file.CustomToolNamespace = value; }
 		}
-		
+
 		protected override bool IsReadOnly (string propertyName)
 		{
 			return false;
-		}
-		
-		[MonoDevelop.Components.PropertyGrid.PropertyEditors.StandardValuesSeparator ("--")]
-		class BuildActionStringsConverter : StandardStringsConverter
-		{
-			public override System.Collections.ICollection GetStandardStrings (ITypeDescriptorContext context)
-			{
-				ProjectFileDescriptor descriptor = context != null?
-					context.Instance as ProjectFileDescriptor : null;
-				
-				if (descriptor != null && descriptor.file != null && descriptor.file.Project != null) {
-					return descriptor.file.Project.GetBuildActions (descriptor.file.FilePath);
-				} else {
-					return new string[] {"Content", "None", "Compile"};
-				}
-			}
-			
-			public override bool CanConvertTo (ITypeDescriptorContext context, Type destinationType)
-			{
-				return destinationType == typeof (string);
-			}
-			
-			public override object ConvertTo (ITypeDescriptorContext context, System.Globalization.CultureInfo culture, object value, Type destinationType)
-			{
-				return (string)value;
-			}
-
-			public override bool CanConvertFrom (ITypeDescriptorContext context, Type sourceType)
-			{
-				return sourceType == typeof (string);
-			}
-			
-			public override object ConvertFrom (ITypeDescriptorContext context,
-			                                    System.Globalization.CultureInfo culture, object value)
-			{
-				if (!IsValid (context, value))
-					throw new FormatException ("Invalid build target name");
-				
-				return (string)value;
-			}
-			
-			public override bool IsValid (ITypeDescriptorContext context, object value)
-			{
-				if (!(value is string))
-					return false;
-				
-				string str = (string) value;
-				if (string.IsNullOrEmpty (str) || !char.IsLetter (str[0]))
-					return false;
-				
-				for (int i = 1; i < str.Length; i++) {
-					char c = str[i];
-					if (char.IsLetterOrDigit (c) || c == '_')
-						continue;
-					else
-						return false;
-				}
-				
-				return true;
-			}
-			public override bool GetStandardValuesExclusive (ITypeDescriptorContext context)
-			{
-				return false;
-			}
-		}
-		
-		abstract class StandardStringsConverter : TypeConverter
-		{
-			public override bool GetStandardValuesSupported (ITypeDescriptorContext context)
-			{
-				return true;
-			}
-		
-			public override bool CanConvertFrom (ITypeDescriptorContext context, Type sourceType)
-			{
-				return sourceType == typeof (string) || base.CanConvertFrom (context, sourceType);
-			}
-		
-			public override bool CanConvertTo (ITypeDescriptorContext context, Type destinationType)
-			{
-				return destinationType == typeof (string) || base.CanConvertTo (context, destinationType);
-			}
-		
-			public override object ConvertFrom (ITypeDescriptorContext context, 
-			                                    System.Globalization.CultureInfo culture, object value)
-			{
-				if (value != null && value is string)
-					return value;
-				else
-					return base.ConvertFrom (context, culture, value);
-			}
-		
-			public override object ConvertTo (ITypeDescriptorContext context, System.Globalization.CultureInfo culture,
-			                                  object value, Type destinationType)
-			{
-				if (value != null && (destinationType == typeof (string)))
-					return value;
-				else
-					return base.ConvertTo (context, culture, value, destinationType);
-			}
-		
-			public override StandardValuesCollection GetStandardValues (ITypeDescriptorContext context)
-			{
-				return new StandardValuesCollection (GetStandardStrings (context));
-			}
-		
-			public abstract System.Collections.ICollection GetStandardStrings (ITypeDescriptorContext context);
 		}
 	}
 }
